Introduction of Style Prototype to extra/card_fields.txt, Shift to complete variable file (arranged by prototype), write and adopt standard switch functions
Introduced the Style Prototype to the extra/card_fields.txt in magic.mse-game. This will be used to switch between styles with different values without needing to go to the style page.

Removed the standard variables file (which held only the variables need for the most basic M15 style) and the Complete_standard_variables which used the card field as the top collection. Both are replaced by complete_variables which uses 'style prototype' as the top collection (complete variables still requires data entry).

Created switch functions to swap between prototypes (in file currently named new document).

Replaced all previous variables and scripts in the style sheet with a basic call of the switch function. (Some values should be dynamic are currently static (and therefore incorrect.)).